size

+++ 0x00-hello_world/6-size.c
@@ -12,10 +12,10 @@ int main(void)
 	char d;
 	float f;
 
-	printf("size of a char: %i byte(s)\n", sizeof(d));
-	printf("size of an int: %i byte(s)\n", sizeof(a));
-	printf("size of a long int: %i byte(s)\n", sizeof(b));
-	printf("size of a long long int: %i byte(s)\n", sizeof(c));
-	printf("size of a float: %i byte(s)\n", sizeof(f)); 	
+	printf("size of a char: %lu byte(s)\n", (unsigned long)sizeof(d));
+	printf("size of an int: %lu byte(s)\n", (unsigned long)sizeof(a));
+	printf("size of a long int: %lu byte(s)\n", (unsigned long)sizeof(b));
+	printf("size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(c));
+	printf("size of a float: %lu byte(s)\n", (unsigned long)sizeof(f)); 	
 	return (0);
 }
